Day 15
this was just a pain in the neck. I tried to use a flat array and the
width of the array versus modeling x/y coords and it caused more
confution than it should have. It did make it easy to print the map but
it made the next steps method confusing