Day 22
Part one was fairly straight forward.
Part two I needed to do some reading on BFS, Dijkstra's algorithm, and
the A* Search algotrithim. Even with that it was helpful to look at some
of the solutions on reddit to help be understand how to apply the
algorithims to this specific problem