Day 10
This one was ... fun. Part two was just counting paths through the tree
but it made my brain hurt. I sketched it out and realized that you could
sum the # of paths from the previous 3 adapters that might be valid. But
it took me forever to get the sum/cache working correctly